Add files via upload
Made the elevator work! It's now capable of opening and closing the doors on its own by using JavaScript. As well, it can hop to floors. It's a bit jumpy at the moment, I need to find a way to make it smoothly move, but outsite of that, it's coming along rather well.